refactor(codec): TextLocation::subslice + replace_range takes Range + html_encode split
- Move `lift_identity(chunk, range)` to `TextLocation::subslice(range)`
  in nvisy-core. The translation is fundamentally about a location
  computing a sub-location; the chunk wrapper was incidental.
  TXT, HTML, PDF, DOCX `lift_chunk` impls become
  `chunk.location.subslice(value_range)`.
- `replace_range` now takes a single `Range<usize>` instead of
  separate `start, end` parameters. Caller sites in CSV, JSON, TXT,
  HTML, and PDF updated. `redact.rs` stays at
  `handler::text::redact::replace_range` — the byte-mutation helper
  is text-family-shaped even though CSV reaches across; renaming
  it earned nothing.
- Split `html_handler.rs` along the natural seam: `EncodePlan` +
  per-kind DOM write helpers move to a new `html_encode` module,
  exposing a single `EncodePlan::apply(&self, &mut Html)` method.
  `html_handler.rs` keeps the handler struct, the `Handler<Text>`
  impl, the inherent methods, and the offset-table helper. ~150 LOC
  out of the handler file.
