+++ crates/nvisy-codec/src/handler/text/mod.rs
@@ -8,38 +8,10 @@
 //! [`Handler::redact`]: crate::Handler::redact
 //! [`TextReplacement`]: nvisy_core::redaction::TextReplacement
 
-use std::ops::Range;
-
-use nvisy_core::modality::{Text, TextLocation};
-
-use crate::Chunk;
-
-/// Identity lift for handlers where `chunk.data` is byte-for-byte
-/// a slice of the source covered by `chunk.location` — no
-/// escapes, no decoding. The default implementation backing
-/// [`Handler::lift_chunk`] for TXT lines, HTML text nodes,
-/// PDF page text, and DOCX text runs.
-///
-/// Adds `value_range` to `chunk.location.start` and copies
-/// `page_number` / `context` through. Returns `None` when the
-/// range falls outside the chunk's value.
-///
-/// [`Handler::lift_chunk`]: crate::Handler::lift_chunk
-pub fn lift_identity(chunk: &Chunk<Text>, value_range: Range<usize>) -> Option<TextLocation> {
-    let len = chunk.location.end.checked_sub(chunk.location.start)?;
-    if value_range.start > value_range.end || value_range.end > len {
-        return None;
-    }
-    Some(TextLocation {
-        start: chunk.location.start + value_range.start,
-        end: chunk.location.start + value_range.end,
-        context: chunk.location.context,
-        page_number: chunk.location.page_number,
-    })
-}
-
 pub(crate) mod redact;
 
+#[cfg(feature = "html")]
+mod html_encode;
 #[cfg(feature = "html")]
 mod html_handler;
 #[cfg(feature = "html")]
